+++ spring-boot-hello-world/src/main/java/com/jalasoft/convert/exception/MyFileNotFoundException.java
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2022 Jala University.
+ *
+ * This software is the confidential and proprieraty information of Jalasoft
+ * ("Confidential Information"). You shall not disclose such Confidential
+ * Information and shall use it only in accordance with the terms of the
+ * Licence agreement you entered into with Jalasoft
+ */
+package com.jalasoft.convert.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class MyFileNotFoundException extends RuntimeException {
+    public MyFileNotFoundException(String message) {
+        super(message);
+    }
+
+    public MyFileNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
