--- spring-boot-hello-world/src/main/java/com/jalasoft/springboothelloworld/SpringBootHelloWorldApplication.java
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) 2022 Jala University.
- *
- * This software is the confidential and proprieraty information of Jalasoft
- * ("Confidential Information"). You shall not disclose such Confidential
- * Information and shall use it only in accordance with the terms of the
- * Licence agreement you entered into with Jalasoft
- */
-package com.jalasoft.springboothelloworld;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringBootHelloWorldApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringBootHelloWorldApplication.class, args);
-	}
-
-}
