+++ spring-boot-hello-world/src/main/resources/application.properties
@@ -1,3 +1,4 @@
+server.port=5000
 #Enable multipart uploads
 spring.servlet.multipart.enabled=true
 # Threshold after which files are written to disk.
